Struct pom::parser::Parser

pub struct Parser<'a, I, O> { /* fields omitted */ }

Parser combinator.

Implementations

impl<'a, I, O> Parser<'a, I, O>

pub fn new<P>(parse: P) -> Parser<'a, I, O> where
P: Fn(&mut dyn Input<I>) -> Result<O> + 'a,

Create new parser.

pub fn parse(&self, input: &mut dyn Input<I>) -> Result<O>

Apply the parser to parse input.

pub fn map<U, F>(self, f: F) -> Parser<'a, I, U> where
F: Fn(O) -> U + 'a,
I: 'static,
O: 'static,
U: 'static,

Convert parser result to desired value.

pub fn convert<U, E, F>(self, f: F) -> Parser<'a, I, U> where
F: Fn(O) -> Result<U, E> + 'a,
E: Debug,
I: Copy + 'static,
O: 'static,
U: 'static,

Convert parser result to desired value, fail in case of conversion error.

pub fn pos(self) -> Parser<'a, I, usize> where
I: Copy + 'static,
O: 'static,

Get input position after matching parser.

pub fn collect(self) -> Parser<'a, I, Vec<I>> where
I: Copy + 'static,
O: 'static,

Collect all matched input symbols.

pub fn discard(self) -> Parser<'a, I, ()> where
I: 'static,
O: 'static,

Discard parser output.

pub fn opt(self) -> Parser<'a, I, Option<O>> where
I: 'static,
O: 'static,

Make parser optional.

pub fn repeat<R>(self, range: R) -> Parser<'a, I, Vec<O>> where
R: RangeArgument<usize> + Debug + 'a,
I: Copy + 'static,
O: 'static,

p.repeat(5) repeat p exactly 5 times p.repeat(0..) repeat p zero or more times p.repeat(1..) repeat p one or more times p.repeat(1..4) match p at least 1 and at most 3 times

pub fn name(self, name: &'a str) -> Parser<'a, I, O> where
I: Copy + 'static,
O: 'static,

Give parser a name to identify parsing errors.

pub fn expect(self, name: &'a str) -> Parser<'a, I, O> where
I: Copy + 'static,
O: 'static,

Mark parser as expected, abort early when failed in ordered choice.