Type Alias pom::Parser

pub type Parser<I, O> = Parser<'static, I, O>;

Parser type, Parser<I, O> is alias of parser::Parser<'static, I, O>.

Aliased Type

struct Parser<I, O> {
    pub method: Box<dyn Fn(&'static [I], usize) -> Result<(O, usize), Error>, Global>,
}

Fields
method: Box<dyn Fn(&'static [I], usize) -> Result<(O, usize), Error>, Global>
Implementations

impl<'a, I, O> Parser<'a, I, O>

pub fn new<P>(parse: P) -> Selfwhere P: Fn(&'a [I], usize) -> Result<(O, usize)> + 'a,

Create new parser.

pub fn parse(&self, input: &'a [I]) -> Result<O>

Apply the parser to parse input.

pub fn parse_at(&self, input: &'a [I], start: usize) -> Result<(O, usize)>

Parse input at specified position.

pub fn map<U, F>(self, f: F) -> Parser<'a, I, U>where F: Fn(O) -> U + 'a, I: 'a, O: 'a, U: 'a,

Convert parser result to desired value.

pub fn convert<U, E, F>(self, f: F) -> Parser<'a, I, U>where F: Fn(O) -> Result<U, E> + 'a, E: Debug, O: 'a, U: 'a,

Convert parser result to desired value, fail in case of conversion error.

pub fn cache(self) -> Selfwhere O: Clone + 'a,

Cache parser output result to speed up backtracking.

pub fn pos(self) -> Parser<'a, I, usize>where O: 'a,

Get input position after matching parser.

pub fn collect(self) -> Parser<'a, I, &'a [I]>where O: 'a,

Collect all matched input symbols.

pub fn discard(self) -> Parser<'a, I, ()>where O: 'a,

Discard parser output.

pub fn opt(self) -> Parser<'a, I, Option<O>>where O: 'a,

Make parser optional.

pub fn repeat<R>(self, range: R) -> Parser<'a, I, Vec<O>>where R: RangeArgument<usize> + Debug + 'a, O: 'a,

p.repeat(5) repeat p exactly 5 times p.repeat(0..) repeat p zero or more times p.repeat(1..) repeat p one or more times p.repeat(1..4) match p at least 1 and at most 3 times

pub fn name(self, name: &'a str) -> Selfwhere O: 'a,

Give parser a name to identify parsing errors.

pub fn expect(self, name: &'a str) -> Selfwhere O: 'a,

Mark parser as expected, abort early when failed in ordered choice.