pom

Module utf8

Variants of parser functions specialized for matching UTF-8 strings and returning chars. Method and constructor names/functionality are generally the same as in base parser module.

Structs

Parser

Being wrapped in this struct guarantees that the parser within will only match valid UTF-8 strings.

Functions

any

Match any UTF-8 character.

call

Call a parser factory, can be used to create recursive parsers.

decode

empty

Always succeeds, consume no input.

end

Success when end of input is reached.

is_a

Success when predicate returns true on current input symbol.

list

Parse separated list.

none_of

Success when current input symbol is none of the set.

not_a

Success when predicate returns false on current input symbol.

one_of

Success when current input symbol is one of the set.

seq

Success when sequence of chars matches current input.

skip

Skip n symbols.

skip_bytes

Skip n bytes exactly.

sym

Match specific UTF-8 character.

take

Read n chars.

take_bytes

Read n bytes exactly.