Crate pom

Modules

char_class

Utility functions to recognize char class of byte value.

parser

Contains predefined parsers and combinators.

utf8

Variants of parser functions specialized for matching UTF-8 strings and returning chars. Method and constructor names/functionality are generally the same as in base parser module.

Enums

Error

Parser error.

Type Aliases

Parser

Parser type, Parser<I, O> is alias of parser::Parser<'static, I, O>.

Result

Parser result, Result<O> ia alias of Result<O, pom::Error>.