Crate pom

Modules

char_class
Utility functions to recognize char class of byte value.

parser
Constains predefined parsers and combinators.

range

set

Structs

DataInput
Wrap &u8 or &char as input to parser.

TextInput
Wrap &str as input to parser.

Enums

Error
Parser error.

Traits

Input
Parser input is generic over terminal type, which is usually u8 or char.

Train
A train is a series of knots, each knot can be a terminal symbol. Train is used as argument of seq(), one_of() and none_of(), so that both b“byte string literal“ and “string literal” can be accepted.

Type Definitions

Parser
Parser type, Parser<I, O> is alias of parser::Parser<'static, I, O>.

Result
Parser result, Result<O> ia alias of Result<O, pom::Error>.