Module pom::char_class

Utility functions to recognize char class of byte value.

Functions

alpha
Recognises an alphabetic character, a-zA-Z.

alphanum
Recognises an alphanumeric character, a-zA-Z0-9.

digit
Recognises a decimal digit, 0-9.

hex_digit
Recognises a hexadecimal digit, 0-9a-fA-F.

multispace
Recognises a space, tab, line feed, or carriage return.

oct_digit
Recognises an octal digit, 0-7.

space
Recognises a space or tab.