pom
Module parser

Constains predefined parsers and combinators.

Structs

Parser

Parser combinator.

Functions

any

Match any symbol.

call

Call a parser factory, can be used to create recursive parsers.

empty

Always succeeds, consume no input.

end

Success when end of input is reached.

is_a

Success when predicate returns true on current input symbol.

list

Parse separated list.

none_of

Success when current input symbol is none of the set.

not_a

Success when predicate returns false on current input symbol.

one_of

Success when current input symbol is one of the set.

seq

Success when sequence of symbols matches current input.

skip

Skip n symbols.

sym

Success when current input symbol equals t.

tag

Success when tag matches current input.

take

Read n symbols.