Struct pom::utf8::Parser

pub struct Parser<'a, O>(/* private fields */);

Being wrapped in this struct guarantees that the parser within will only match valid UTF-8 strings.

Implementations

impl<'a, O> Parser<'a, O>

pub fn new<P>(parse: P) -> Self
where P: Fn(&'a [u8], usize) -> Result<(O, usize)> + 'a,

Create new parser.

pub fn collect(self) -> Parser<'a, &'a str>
where O: 'a,

Collect all matched input symbols.

pub fn parse(&self, input: &'a [u8]) -> Result<O>

Apply the parser to parse input.

pub fn parse_at(&self, input: &'a [u8], start: usize) -> Result<(O, usize)>

Parse input at specified byte position.

pub fn parse_str(&self, input: &'a str) -> Result<O>

Apply the parser to parse input.

pub fn map<U, F>(self, f: F) -> Parser<'a, U>
where F: Fn(O) -> U + 'a, O: 'a, U: 'a,

Convert parser result to desired value.

pub fn convert<U, E, F>(self, f: F) -> Parser<'a, U>
where F: Fn(O) -> Result<U, E> + 'a, E: Debug, O: 'a, U: 'a,

Convert parser result to desired value, fail in case of conversion error.

pub fn cache(self) -> Self
where O: Clone + 'a,

Cache parser output result to speed up backtracking.

pub fn pos(self) -> Parser<'a, usize>
where O: 'a,

Get input position after matching parser.

pub fn discard(self) -> Parser<'a, ()>
where O: 'a,

Discard parser output.

pub fn opt(self) -> Parser<'a, Option<O>>
where O: 'a,

Make parser optional.

pub fn repeat<R>(self, range: R) -> Parser<'a, Vec<O>>
where R: RangeArgument<usize> + Debug + 'a, O: 'a,

p.repeat(5) repeat p exactly 5 times p.repeat(0..) repeat p zero or more times p.repeat(1..) repeat p one or more times p.repeat(1..4) match p at least 1 and at most 3 times

pub fn name(self, name: &'a str) -> Self
where O: 'a,

Give parser a name to identify parsing errors.

pub fn expect(self, name: &'a str) -> Self
where O: 'a,

Mark parser as expected, abort early when failed in ordered choice.